requirements.txt
# Web服务与API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 网页转PDF
playwright>=1.40.0

# 异步任务队列
celery>=5.3.0
redis>=5.0.0

# 文档处理
python-docx>=1.1.0
beautifulsoup4>=4.12.0

# 日志系统
loguru>=0.7.0

# 邮件服务
python-multipart>=0.0.6

# 数据库（为后台管理准备）
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# 开发工具
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# 新增依赖
email-validator>=2.0.0

# 微信机器人
wechaty>=0.9.0
wechaty-puppet-service>=0.3.0
